refactor: rename registry module to cleanup, move ShmemKind to lib.rs
The old 'registry' name was a leftover from the removed ShmemRegistry.
- ShmemKind is a crate-level enum, now lives in lib.rs
- cleanup_flink, cleanup_shmem, is_pid_alive move to cleanup.rs
- All downstream imports updated to use top-level re-exports
