+++ crates/flux-ctl/src/discovery/mod.rs
@@ -5,7 +5,7 @@ pub mod inspect;
 pub mod registry;
 
 pub use cli::{clean, inspect, list_all, list_json, stats};
-pub use flux_communication::registry::is_pid_alive;
+pub use flux_communication::is_pid_alive;
 pub use inspect::{
     PidInfo, PoisonInfo, QueueStats, backing_file_size, format_bytes, scan_proc_fds,
 };
