+++ crates/flux-network/src/tcp/connector.rs
@@ -182,8 +182,8 @@ impl ConnectionManager {
         if let Some(stream) = self.try_connect(o, addr) {
             let mut tcp_stream =
                 TcpStream::from_stream_with_telemetry(stream, o, addr, self.telemetry);
-            if let Some(msg) = &self.on_connect_msg
-                && tcp_stream.write_or_enqueue_with(self.poll.registry(), |buf: &mut Vec<u8>| {
+            if let Some(msg) = &self.on_connect_msg &&
+                tcp_stream.write_or_enqueue_with(self.poll.registry(), |buf: &mut Vec<u8>| {
                     buf.extend_from_slice(msg);
                 }) == ConnState::Disconnected
             {
